Modified token auth code to support python 2. (#41)
